+++ sporttracker/gpx/LongDistanceTourGpxPreviewImageService.py
@@ -14,7 +14,6 @@ import requests
 from sporttracker import Constants
 from sporttracker.longDistanceTour.LongDistanceTourEntity import LongDistanceTour
 
-
 LOGGER = logging.getLogger(Constants.APP_NAME)
 
 
@@ -81,7 +80,8 @@ class LongDistanceTourGpxPreviewImageService:
 
                 with open(tempGpxFilePath, 'rb') as fd:
                     files = {'gpx': fd}
-                    response = requests.post(gpxPreviewImageSettings['geoRenderUrl'], files=files)
+                    timeout = gpxPreviewImageSettings['timeout']
+                    response = requests.post(gpxPreviewImageSettings['geoRenderUrl'], files=files, timeout=timeout)
                     response.raise_for_status()
 
                     with open(self.get_preview_image_path(), 'wb') as f: